pixel width revisited

+++ api/src/test/java/net/kyori/adventure/text/width/CustomFontCharacterWidthFunction.java
@@ -29,12 +29,12 @@ import org.jetbrains.annotations.NotNull;
 
 public class CustomFontCharacterWidthFunction implements CharacterWidthFunction {
   @Override
-  public int widthOf(final int codepoint, final @NotNull Style style) {
-    if(Character.isLowerCase(codepoint)) return 3;
-    if(Character.isUpperCase(codepoint)) return 5;
-    if(Character.isDigit(codepoint)) return style.hasDecoration(TextDecoration.OBFUSCATED) ? 4 : 3;
-    if(Character.isSpaceChar(codepoint)) return 2;
-    if(codepoint == 65938) return 8; //𐆒
+  public float widthOf(final int codepoint, final @NotNull Style style) {
+    if (Character.isLowerCase(codepoint)) return 3;
+    if (Character.isUpperCase(codepoint)) return 5;
+    if (Character.isDigit(codepoint)) return style.hasDecoration(TextDecoration.OBFUSCATED) ? 4 : 3;
+    if (Character.isSpaceChar(codepoint)) return 2;
+    if (codepoint == 65938) return 8; //𐆒
     return 0;
   }
 }
